+++ Calorie Scanner Tracker And Diet Suggestor/Models/Meals.cs
@@ -8,10 +8,10 @@ namespace Calorie_Scanner_Tracker_And_Diet_Suggestor.Models
         public int Id { get; set; }
         [Required]
         public string? Name { get; set; }
-        public int Calories { get; set; }
-        public decimal Protein { get; set; }
-        public decimal Carbs { get; set; }
-        public decimal Fats { get; set; }
+        public decimal Calories { get; set; } // Change to double
+        public decimal Protein { get; set; }  // Change to double
+        public decimal Carbs { get; set; }    // Change to double
+        public decimal Fats { get; set; }     // Change to double
         public string? MealType { get; set; } // Breakfast, Lunch, Dinner
 
         public List<PreparationStep> PreparationSteps { get; set; } = new List<PreparationStep>();
